requirements.txt
fastapi
uvicorn
httpx
pydantic>=2
python-dotenv
psycopg[binary]
redis
tenacity